refactor[GetCapaDoPassiva]
